requirements.txt
pandas  
pymongo  
Flask-SQLAlchemy  
textblob
finnhub-python